requirements.txt
accelerate==0.25.0
aiohttp==3.9.1
aiosignal==1.3.1
angle-emb==0.1.6
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
boltons==23.1.1
bottle==0.12.25
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
CProfileV==1.0.7
cycler==0.12.1
datasets==2.16.1
dill==0.3.7
faiss-cpu==1.7.4
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.10.0
gdown==5.0.1
huggingface-hub==0.20.2
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
line-profiler==4.1.2
lxml==5.1.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
packaging==23.2
pandas==2.1.4
peft==0.7.1
pillow==10.2.0
prettytable==3.9.0
psutil==5.9.7
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
sarge==0.1.7.post1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
soupsieve==2.5
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
trectools==0.0.49
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
xxhash==3.4.1
yarl==1.9.4
